# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Monday API Automation",
    "description": "API para automatizar a criação de estrutura de pastas baseada em demandas do Monday.com, definindo analistas responsáveis por produto.",
    "version": "1.0.0",
    "contact": {
      "name": "Nathan Silva - Fagron Tech",
      "email": "nathan.silva@fagrontech.com.br"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "https://monday-api-automation.onrender.com",
      "description": "Servidor de Produção (Render) - GitHub: https://github.com/Nathan-Paranhos/monday-api-automation"
    },
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desenvolvimento Local"
    }
  ],
  "paths": {
...
    "/api/monday/products/{clientId}": {
      "get": {
        "tags": ["Monday.com"],
        "summary": "Consultar Produto por ID",
        "description": "Busca o produto associado a um cliente específico no Monday.com",
        "operationId": "queryProduct",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "description": "ID do cliente",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProdutoResponse": {
        "type": "object",
        "properties": {
          "operacao_status": {
            "type": "string",
            "enum": ["ok"],
            "description": "Status da consulta"
          },
          "id_cliente": {
            "type": "integer",
            "description": "ID do cliente consultado"
          },
          "produto": {
            "type": "string",
            "enum": ["BOT", "Fórmula Certa", "Phusion"],
            "description": "Produto encontrado"
          },
          "principal_produto": {
            "type": "string",
            "description": "Principal produto associado ao item no Monday.com"
          },
          "status": {
            "type": "object",
            "properties": {
              "original": {
                "type": "string",
                "description": "Status original do Monday.com"
              },
              "normalizado": {
                "type": "string",
                "description": "Status normalizado para uso interno"
              }
            },
            "description": "Informações de status do item"
          },
          "elemento": {
            "type": "string",
            "description": "Campo elemento do Monday.com que contém o código do cliente e nome da farmácia (formato: 'CÓDIGO - NOME_FARMACIA')",
            "example": "12345 - Farmácia Exemplo"
          },
          "campos": {
            "type": "object",
            "additionalProperties": true,
            "description": "Campos adicionais do Monday.com"
          },
          "responsavel": {
            "type": "string",
            "description": "Email do responsável pelo produto. Para o produto BOT, contém múltiplos emails separados por vírgula."
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp da consulta"
          }
        }
      },
...
    }
  },
...
}